Turn `ContractsFromMethod` into an interface and add `NoContractsFromMethod` implementation
