+++ framework/src/main/java/org/checkerframework/common/initializedfields/InitializedFieldsAnnotatedTypeFactory.java
@@ -15,7 +15,7 @@ import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
 import org.checkerframework.framework.util.Contract;
-import org.checkerframework.framework.util.ContractsFromMethod;
+import org.checkerframework.framework.util.DefaultContractsFromMethod;
 import org.checkerframework.javacutil.AnnotationBuilder;
 import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.UserError;
@@ -121,7 +121,7 @@ public class InitializedFieldsAnnotatedTypeFactory extends AccumulationAnnotated
      * A subclass of ContractsFromMethod that adds a postcondition contract to each constructor,
      * requiring that it initializes all fields.
      */
-    private class InitializedFieldsContractsFromMethod extends ContractsFromMethod {
+    private class InitializedFieldsContractsFromMethod extends DefaultContractsFromMethod {
         /**
          * Creates an InitializedFieldsContractsFromMethod for the given factory.
          *
@@ -136,39 +136,31 @@ public class InitializedFieldsAnnotatedTypeFactory extends AccumulationAnnotated
         public Set<Contract.Postcondition> getPostconditions(ExecutableElement executableElement) {
             Set<Contract.Postcondition> result = super.getPostconditions(executableElement);
 
-            // Only process methods defined in source code being type-checked.
-            if (declarationFromElement(executableElement) != null) {
-
-                if (executableElement.getKind() == ElementKind.CONSTRUCTOR) {
-                    // It's a constructor
-
-                    String[] fieldsToInitialize =
-                            fieldsToInitialize(
-                                    (TypeElement) executableElement.getEnclosingElement());
-                    if (fieldsToInitialize.length != 0) {
-
-                        AnnotationMirror initializedFieldsAnno;
-                        {
-                            AnnotationBuilder builder =
-                                    new AnnotationBuilder(processingEnv, InitializedFields.class);
-                            builder.setValue("value", fieldsToInitialize);
-                            initializedFieldsAnno = builder.build();
-                        }
-                        AnnotationMirror ensuresAnno;
-                        {
-                            AnnotationBuilder builder =
-                                    new AnnotationBuilder(
-                                            processingEnv, EnsuresInitializedFields.class);
-                            builder.setValue("value", thisStringArray);
-                            builder.setValue("fields", fieldsToInitialize);
-                            ensuresAnno = builder.build();
-                        }
-                        Contract.Postcondition ensuresContract =
-                                new Contract.Postcondition(
-                                        "this", initializedFieldsAnno, ensuresAnno);
-
-                        result.add(ensuresContract);
+            // Only process constructors defined in source code being type-checked.
+            if (declarationFromElement(executableElement) != null
+                    && executableElement.getKind() == ElementKind.CONSTRUCTOR) {
+                String[] fieldsToInitialize =
+                        fieldsToInitialize((TypeElement) executableElement.getEnclosingElement());
+                if (fieldsToInitialize.length != 0) {
+                    AnnotationMirror initializedFieldsAnno;
+                    {
+                        AnnotationBuilder builder =
+                                new AnnotationBuilder(processingEnv, InitializedFields.class);
+                        builder.setValue("value", fieldsToInitialize);
+                        initializedFieldsAnno = builder.build();
                     }
+                    AnnotationMirror ensuresAnno;
+                    {
+                        AnnotationBuilder builder =
+                                new AnnotationBuilder(
+                                        processingEnv, EnsuresInitializedFields.class);
+                        builder.setValue("value", thisStringArray);
+                        builder.setValue("fields", fieldsToInitialize);
+                        ensuresAnno = builder.build();
+                    }
+                    Contract.Postcondition ensuresContract =
+                            new Contract.Postcondition("this", initializedFieldsAnno, ensuresAnno);
+                    result.add(ensuresContract);
                 }
             }
